+++ packages/block-library/src/cover/edit/resizeable-cover.js
